ugly ply python pythagoras godzilla python pythagoras godzilla
